refactor: minimize public surface, expose operator_tuples and Tuple
Reduces the library's externally reachable API by making the top-level
module declarations private to the crate. Items within those modules
keep plain `pub` since Clippy's redundant_pub_crate lint flags pub(crate)
inside an already-private module as redundant.

Changes:
- constraints, solver, geometry in lib.rs: pub mod → pub(crate) mod
- operator_tuples promoted to pub and re-exported from lib.rs
- Tuple re-exported from lib.rs (required by operator_tuples signature)

The effect: external users can access only what is re-exported from
lib.rs. All internal module structure is hidden.
