+++ src/constraints/cage/mod.rs
@@ -1,3 +1,5 @@
 pub mod arithmetic;
 pub mod builder;
 pub mod operation;
+
+pub use builder::operator_tuples;
