+++ src/constraints/mod.rs
@@ -1,0 +1,4 @@
+pub mod all_different;
+pub mod cage;
+pub mod constraint;
+mod regin;
