+++ src/generation/latin_square.rs
