+++ src/generation/mod.rs
@@ -1,0 +1,4 @@
+pub mod generate;
+mod latin_square;
+
+pub use generate::generate;
