+++ src/geometry/grid.rs
