+++ src/geometry/mod.rs
@@ -1,0 +1,3 @@
+pub mod grid;
+pub mod shape;
+pub mod tiling;
