+++ src/lib.rs
@@ -1,23 +1,18 @@
-pub mod all_different;
-mod cage;
-mod constraint;
-mod delta;
+pub(crate) mod constraints;
 mod generation;
-mod grid;
-mod latin_square;
+pub(crate) mod geometry;
 mod puzzle;
-mod regin;
-mod shape;
-mod solver;
-mod tiling;
+pub(crate) mod solver;
 mod types;
 
-pub use cage::builder::Cage;
-pub use cage::operation::{Operation, Operator};
-pub use delta::Delta;
-pub use generation::{DEFAULT_SIZE_DISTRIBUTION, default_op_policy, generate, generate_with};
-pub use grid::Grid;
+pub use constraints::cage::builder::{Cage, Tuple};
+pub use constraints::cage::operation::{Operation, Operator};
+pub use constraints::cage::operator_tuples;
+pub use generation::generate;
+pub use generation::generate::{DEFAULT_SIZE_DISTRIBUTION, default_op_policy, generate_with};
+pub use geometry::grid::Grid;
+pub use geometry::shape::Polyomino;
+pub use geometry::tiling::SizeDistribution;
 pub use puzzle::{NarrowingScore, Puzzle, Uniqueness};
-pub use shape::Polyomino;
-pub use tiling::SizeDistribution;
+pub use solver::delta::Delta;
 pub use types::{Cell, Error, Index, M, N, Values};
