+++ src/solver/solve.rs
